+++ 03_Modules/Certificates/src/index.ts
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { CertificatesOcpp201Api } from './module/2.0.1/api';
-export { CertificatesDataApi } from './module/api';
+export { CertificatesOcpp201Api } from './module/2.0.1/MessageApi';
+export { CertificatesDataApi } from './module/DataApi';
 export { ICertificatesModuleApi } from './module/interface';
 export { CertificatesModule } from './module/module';
