+++ 03_Modules/Configuration/src/index.ts
@@ -3,8 +3,8 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { ConfigurationDataApi } from './module/api';
-export { ConfigurationOcpp201Api } from './module/2.0.1/api';
-export { ConfigurationOcpp16Api } from './module/1.6/api';
+export { ConfigurationDataApi } from './module/DataApi';
+export { ConfigurationOcpp201Api } from './module/2.0.1/MessageApi';
+export { ConfigurationOcpp16Api } from './module/1.6/MessageApi';
 export { IConfigurationModuleApi } from './module/interface';
 export { ConfigurationModule } from './module/module';
