+++ 03_Modules/EVDriver/src/index.ts
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { EVDriverOcpp201Api } from './module/2.0.1/api';
-export { EVDriverDataApi } from './module/api';
+export { EVDriverOcpp201Api } from './module/2.0.1/MessageApi';
+export { EVDriverDataApi } from './module/DataApi';
 export { IEVDriverModuleApi } from './module/interface';
 export { EVDriverModule } from './module/module';
