+++ 03_Modules/SmartCharging/src/index.ts
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { SmartChargingOcpp201Api } from './module/2.0.1/api';
+export { SmartChargingOcpp201Api } from './module/2.0.1/MessageApi';
 export { ISmartChargingModuleApi } from './module/interface';
 export { SmartChargingModule } from './module/module';
 export { ISmartCharging, InternalSmartCharging } from './module/smartCharging';
