+++ 03_Modules/Transactions/src/index.ts
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { TransactionsOcpp201Api } from './module/2.0.1/api';
-export { TransactionsDataApi } from './module/api';
+export { TransactionsOcpp201Api } from './module/2.0.1/MessageApi';
+export { TransactionsDataApi } from './module/DataApi';
 export { ITransactionsModuleApi } from './module/interface';
 export { TransactionsModule } from './module/module';
